backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
python-dotenv
pandas
openpyxl
